\name{expp-package}
\alias{expp-package}
\alias{expp}
\docType{package}
\title{
Tools and data to accompany Schlicht, Valcu and Kempenaers 
"Spatial patterns of extra-pair paternity: beyond paternity gain and loss"
}

\description{
The expp package provides  classes and functions for the investigation of the probability of having extra-pair young within local networks of breeding pairs including both realized and potential extra-pairings. 
}

\details{
\tabular{ll}{
Package: \tab expp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-22\cr
License: \tab GPL-3\cr
}
See \code{help(epp)} and 
\code{vignette('expp') }
}

\author{
 Mihai Valcu and Lotte Schlicht
Maintainer: Mihai Valcu <valcu@orn.mpg.de>

}
\references{
Schlicht, Valcu and Kempenaers 
"Spatial patterns of extra-pair paternity: beyond paternity gain and loss" (in prep.)
}

\keyword{ package }

