% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expp-package.R
\docType{data}
\name{bluetit_boundary}
\alias{bluetit_boundary}
\title{Study area boundary.}
\format{A SpatialPolygonsDataFrame with 12 SpatialPolygons.  
\describe{
 \item{list("year_")}{numeric. The year of the observation.} 
 }}
\description{
\code{SpatialPolygonsDataFrame} Study area boundary of two Blue Tit
populations: Kolbeterberg, Vienna, Austria (1998 through 2004) and
Westerholz, Bavaria, Germany (2007 through 2011) .
}
\examples{

data(bluetit_boundary)
summary(bluetit_boundary)

}
\keyword{datasets}
