% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expp-package.R
\docType{data}
\name{bluetit_breeding}
\alias{bluetit_breeding}
\title{Blue Tit breeding data.}
\format{A data frame with 1025 observations on the following 10 variables.
\itemize{ 
     \item year_       numeric. The year of the observation.
     \item id          numeric. The identity of the nest box in which the
                       breeding attempt took place. 
     \item x           numeric. The east-west location of the nest box. 
     \item y           numeric. The north-south location of the nest box. 
     \item female      character. The identity of the female. 
     \item male        character. The identity of the male. 
     \item layingDate  numeric. The day of the year when the first egg was produced. 
     \item male_age    character. The age class of the male ('juv' = 1st year breeder; 'adult' = older)
     \item male_tarsus numeric. tarsus length (mm))
     \item study_area  character. The study area name.  
     }}
\description{
Breeding data recorded for two Blue Tit populations in Kolbeterberg, Vienna,
Austria (1998 through 2004) and Westerholz, Bavaria, Germany (2007 through
2011) . The data set contains breeding attempts locations, the respective
social pair, and several individual and nest parameters.
}
\examples{

data(bluetit_breeding)
head(bluetit_breeding)

}
\keyword{datasets}
