% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expp-package.R
\docType{data}
\name{bluetit_epp}
\alias{bluetit_epp}
\title{Blue tit extra-pair paternity data.}
\format{A data frame with 425 observations on the following 3 variables.
\itemize{ 
     \item year_   numeric. The year of the observation.
     \item female   character. The female involved in the respective EPP event.
     \item male     character. The male involved in the respective EPP event.
     }}
\description{
\code{data.frame} Extra-pair paternity data recorded for two Blue Tit
populations in Kolbeterberg, Vienna, Austria (1998 through 2004) and
Westerholz, Bavaria, Germany (2007 through 2011) .
}
\examples{

data(bluetit_epp)
head(bluetit_epp)

}
\keyword{datasets}
